advanced feedback and favourite feedback add average star and number of Feedback for Product change delete and write feedback

+++ src/App/controllers/FeedbackController/index.js
@@ -2,6 +2,7 @@ const  mongoose  = require('mongoose');
 const FavouriteFeedback = require('../../models/FavouriteFeedback')
 const Feedback  = require('../../models/Feedback');
 const isValidID = require('../../../utils/isValidID');
+const Product = require('../../models/Product');
 
 const LIMIT = 5;
 
@@ -47,16 +48,28 @@ class FeedbackController{
     // }
     async write(req, res, next) {
         try {
+            const stars = Number(req.body.stars)
             const userID = isValidID(req.body.userID)
             const idProduct = isValidID(req.body.idProduct)
-            const comment = req.body?.comment|| ''
-            const data = await Feedback.create({userID,idProduct,comment})
-            
-            if(data) res.status(200).json({
-                success:true,
-                title:'success',
-                data
+            const comment = req.body?.comment|| 'Người dùng không viết đánh giá'
+            const product = await Product.findById(idProduct)           
+            if(stars >0 && userID && idProduct && product){
+                const data = await Feedback.create({userID,idProduct,comment,stars})
+                const starAverage = (product.starAverage*product.numberFeedback+stars)/(product.numberFeedback + 1)
+                await Product.updateOne({_id:idProduct},{$inc:{numberFeedback:1},$set:{starAverage:starAverage}})
+                
+                if(data) return res.status(200).json({
+                    success:true,
+                    title:'success',
+                    data
+                })
+            }
+            return res.status(403).json({
+                title:'error',
+                success:false,
+                message:'Check stars , userID and idProduct'
             })
+            
         } catch (error) {
             console.log(error);
         }
@@ -68,20 +81,32 @@ class FeedbackController{
             const userID = isValidID(req.body.userID)
             const idProduct = isValidID(req.body.idProduct)
             const idFeedback = isValidID(req.body.idFeedback)
-
-            const data = await Feedback.deleteOne({userID,idProduct,_id:idFeedback})
-            if(data.acknowledged && data.deletedCount) res.status(200).json({
-                title:'success',
-                success:true,
-                data
-            })
-            else{
-                res.status(403).json({
-                    success:false,
-                    title:'invalid',
+            const product = await Product.findById(idProduct)
+            const feedback = await Feedback.findById(idFeedback)        
+            if(product && feedback){
+                const starAverage = (product?.starAverage*product?.numberFeedback-feedback?.stars)/(product?.numberFeedback - 1)
+                console.log(product,feedback);
+                await Product.updateOne({_id:idProduct},{$inc:{numberFeedback:-1},$set:{starAverage:starAverage}})
+                const data = await Feedback.deleteOne({userID,idProduct,_id:idFeedback})
+                if(data.acknowledged && data.deletedCount) return res.status(200).json({
+                    title:'success',
+                    success:true,
                     data
                 })
-            }
+                else{
+                    return res.status(403).json({
+                        success:false,
+                        title:'invalid',
+                        data
+                    })
+                }
+            }    
+            else return res.status(403).json({
+                title:'error',
+                success:false,
+                message:'checking input'
+            })
+           
         } catch (error) {
             console.log(error);
         }
